requirements.txt
tkintertable
numpy
opencv-python
opencv-contrib-python
tqdm
Pillow
matplotlib
tensorflow
tensorflow-addons
